Fix a typo on the ynVar typedef that unfortunately caused 3.6.20 to be released with a "bug" that causes excess memory usage for Expr objects. This bug should be benign in the sense that the correct answer is still always obtained - it only uses more more memory than is really necessary.
FossilOrigin-Name: f1c09acaca3e205acf5b077c9b2d0fe35f035c1e

+++ src/sqliteInt.h
@@ -1508,7 +1508,7 @@ struct AggInfo {
 ** the option is available (at compile-time).
 */
 #if SQLITE_MAX_VARIABLE_NUMBER<=32767
-typedef i64 ynVar;
+typedef i16 ynVar;
 #else
 typedef int ynVar;
 #endif
